Controllers completados mayoria

+++ src/main/java/com/Backend/MediConnect/market/domain/repository/ConsultaRepository.java
@@ -3,6 +3,4 @@ package com.Backend.MediConnect.market.domain.repository;
 import com.Backend.MediConnect.market.persistance.entity.Consulta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConsultaRepository extends JpaRepository<Consulta, Long> {
-
-}
+public interface ConsultaRepository extends JpaRepository<Consulta, Integer> {}
