+++ src/main/java/com/Backend/MediConnect/market/domain/repository/RecetaRepository.java
@@ -3,6 +3,4 @@ package com.Backend.MediConnect.market.domain.repository;
 import com.Backend.MediConnect.market.persistance.entity.Receta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecetaRepository extends JpaRepository<Receta, Long>{
-
-}
+public interface RecetaRepository extends JpaRepository<Receta, Integer> {}
